Clean up precompiles crate
