+++ precompiles/src/neuron.rs
@@ -8,7 +8,8 @@ use sp_core::H256;
 use sp_runtime::traits::Dispatchable;
 use sp_std::vec::Vec;
 
-use crate::{PrecompileExt, PrecompileHandleExt, parse_pubkey};
+use crate::parser::parse_pubkey;
+use crate::{PrecompileExt, PrecompileHandleExt};
 
 pub struct NeuronPrecompile<R>(PhantomData<R>);
